Create directory for filter module, move median_filter into separate file
